tests/ftests: synchronize between github runners
If a VM is shared between multiple github runners, there are chances of
each other stomping over other's run, if executed parallelly.  To avoid
the race between the runners, introduce lock file, that gets acquired
(created) when ftest.sh starts and get removed by ftest-nocontainer.sh,
this ensures that both test cases are executed before other runner,
that's waiting for its chance to run.  A runner would wait for
10 minutes before re-trying to run.  At the max, a runner would wait for
50 minutes (5 retries) before giving up.

+++ tests/ftests/ftests-nocontainer.sh
@@ -6,6 +6,22 @@ AUTOMAKE_HARD_ERROR=99
 
 START_DIR=$PWD
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
+RUNNER_LOCK_FILE="/var/run/libcgroup/github-runner.lock"
+
+# the lock file is removed after executing no-container
+# test cases, whereas the lock is taken by the ftests.sh
+function cleanup()
+{
+	sudo rm -f "$RUNNER_LOCK_FILE"
+	exit "$1"
+}
+
+# ftests.sh should have taken the lock, if not
+# something is out of order, let's abort
+if [ ! -f "$RUNNER_LOCK_FILE" ]; then
+	echo "Lock file $RUNNER_LOCK_FILE missing, aborting"
+	exit 1
+fi
 
 if [ "$START_DIR" != "$SCRIPT_DIR" ]; then
 	cp "$SCRIPT_DIR"/*.py "$START_DIR"
@@ -21,10 +37,10 @@ fi
 	-n Libcg"$RANDOM"
 RET1=$?
 
-pushd ../../src || exit $AUTOMAKE_HARD_ERROR
+pushd ../../src || cleanup $AUTOMAKE_HARD_ERROR
 PATH="$PATH:$(pwd)"
 export PATH
-popd || exit $AUTOMAKE_HARD_ERROR
+popd || cleanp $AUTOMAKE_HARD_ERROR
 
 sudo PATH=$PATH PYTHONPATH=$PYTHONPATH ./ftests.py -l 10 -s "sudo" \
 	-L "$START_DIR/ftests-nocontainer.py.sudo.log" --no-container -n Libcg"$RANDOM"
@@ -40,20 +56,20 @@ fi
 
 if [[ $RET1 -ne $AUTOMAKE_SKIPPED ]] && [[ $RET1 -ne 0 ]]; then
 	# always return errors from the first test run
-	exit $RET1
+	cleanup $RET1
 fi
 if [[ $RET2 -ne $AUTOMAKE_SKIPPED ]] && [[ $RET2 -ne 0 ]]; then
 	# return errors from the second test run
-	exit $RET2
+	cleanup $RET2
 fi
 
 if [[ $RET1 -eq 0 ]] || [[ $RET2 -eq 0 ]]; then
-	exit 0
+	cleanup 0
 fi
 
 if [[ $RET1 -eq $AUTOMAKE_SKIPPED ]] || [[ $RET2 -eq $AUTOMAKE_SKIPPED ]]; then
-	exit $AUTOMAKE_SKIPPED
+	cleanup $AUTOMAKE_SKIPPED
 fi
 
 # I don't think we should ever get here, but better safe than sorry
-exit $AUTOMAKE_HARD_ERROR
+cleanup $AUTOMAKE_HARD_ERROR
